14-ollamahragapp/requirements.txt
accelerate==1.8.1
ai21==4.0.3
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiolimiter==1.2.1
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anthropic==0.57.1
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
asyncpg==0.30.0
attrs==25.3.0
babel==2.17.0
banks==2.2.0
bce-python-sdk==0.9.35
beautifulsoup4==4.13.4
bitsandbytes==0.46.1
bleach==6.2.0
blis==0.7.11
boto3==1.39.3
botocore==1.39.3
bytecode==0.16.2
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.1
cmake==4.0.3
cohere==5.15.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
cut-cross-entropy==25.1.1
cymem==2.0.11
dataclasses-json==0.6.7
datadog==0.52.1
datadog_lambda==6.111.0
datasets==3.6.0
ddtrace==3.11.2
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
diffusers==0.34.0
dill==0.3.8
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.16
einops==0.8.1
envier==0.6.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
evaluate==0.4.5
executing==2.2.0
faiss-cpu==1.11.0.post1
fastavro==1.11.1
fastjsonschema==2.21.1
filelock==3.18.0
filetype==1.2.0
fire==0.7.0
flash_attn==2.8.0.post2
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.3.0
future==1.0.0
get-settings==0.1.1
gguf==0.17.1
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
gpt4all==2.8.2
greenlet==3.2.3
griffe==1.7.3
grpcio==1.73.1
grpcio-status==1.48.2
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.5
hf_transfer==0.1.9
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.2
idna==3.10
importlib_metadata==8.7.0
inquirerpy==0.3.4
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
json5==0.12.0
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
langchain==0.3.26
langchain-anthropic==0.3.17
langchain-aws==0.2.27
langchain-cohere==0.4.4
langchain-community==0.3.27
langchain-core==0.3.68
langchain-google-genai==2.1.6
langchain-mistralai==0.2.10
langchain-nvidia-ai-endpoints==0.3.10
langchain-ollama==0.3.3
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langcodes==3.5.0
langsmith==0.4.4
language_data==1.3.0
llama-cloud==0.1.32
llama-cloud-services==0.6.43
llama-index==0.12.52
llama-index-agent-openai==0.4.12
llama-index-cli==0.4.4
llama-index-core==0.12.52.post1
llama-index-embeddings-huggingface==0.5.5
llama-index-embeddings-ollama==0.6.0
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.7.10
llama-index-instrumentation==0.2.0
llama-index-llms-llama-cpp==0.4.0
llama-index-llms-ollama==0.6.2
llama-index-llms-openai==0.4.7
llama-index-multi-modal-llms-openai==0.5.3
llama-index-program-openai==0.3.2
llama-index-question-gen-openai==0.3.1
llama-index-readers-file==0.4.11
llama-index-readers-llama-parse==0.4.0
llama-index-vector-stores-faiss==0.4.0
llama-index-vector-stores-postgres==0.5.4
llama-index-workflows==1.0.1
llama-parse==0.6.43
llama_cpp_python==0.3.14
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mergedeep==1.3.4
mistral_common==1.8.1
mistral_inference==1.6.0
mistune==3.1.3
more-itertools==10.7.0
mpmath==1.3.0
msgspec==0.19.0
multidict==6.6.3
multiprocess==0.70.16
murmurhash==1.0.13
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
notebook==7.4.4
notebook_shim==0.2.4
numpy==2.3.1
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
ollama==0.5.1
openai==1.93.0
openpyxl==3.1.5
opentelemetry-api==1.36.0
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
path==17.1.1
peft==0.16.0
pexpect==4.9.0
pfzy==0.3.4
pgai==0.11.4
pgvector==0.4.1
pillow==11.3.0
platformdirs==4.3.8
ply==3.11
preshed==3.0.10
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==3.20.3
psutil==7.0.0
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyMuPDF==1.26.3
pypdf==5.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
qianfan==0.4.12.3
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
s3transfer==0.13.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
semver==3.0.4
Send2Trash==1.8.3
sentence-transformers==5.0.0
sentencepiece==0.2.0
setuptools==80.9.0
shellingham==1.5.4
shtab==1.7.2
simple-parsing==0.1.7
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
soupsieve==2.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
stack-data==0.6.3
striprtf==0.0.26
structlog==25.4.0
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
termcolor==3.1.0
terminado==0.18.1
thefuzz==0.22.1
thinc==8.2.3
threadpoolctl==3.6.0
tiktoken==0.9.0
timescale-vector==0.0.7
tinycss2==1.4.0
together==1.5.17
tokenizers==0.21.2
torch==2.7.0
torchvision==0.22.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
triton==3.3.0
trl==0.19.1
typeguard==4.4.4
typer==0.15.4
types-python-dateutil==2.9.0.20250516
types-PyYAML==6.0.12.20250516
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tyro==0.9.26
tzdata==2025.2
ujson==5.10.0
unsloth @ git+https://github.com/unslothai/unsloth.git@89ba48dd7e5c1d02798d83fa5e347d024ad8c914
unsloth_zoo @ git+https://github.com/unslothai/unsloth_zoo.git@7b7781299dc3c2bddb8f4ab6eb4bde1f75b2a5c9
uri-template==1.3.0
urllib3==2.5.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.45.1
widgetsnbextension==4.0.14
wrapt==1.17.2
xformers==0.0.30
xmltodict==0.14.2
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
